scripts/tests/requirements.txt
# Test Requirements

pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0

# For test fixtures
Pillow>=10.0.0
